Function amazon_sp_product_fees::apis::fees_api::get_my_fees_estimate_for_sku

pub async fn get_my_fees_estimate_for_sku(
    configuration: &Configuration,
    seller_sku: &str,
    body: GetMyFeesEstimateRequest
) -> Result<GetMyFeesEstimateResponse, Error<GetMyFeesEstimateForSkuError>>

Returns the estimated fees for the item indicated by the specified seller SKU in the marketplace specified in the request body. You can call getMyFeesEstimateForSKU for an item on behalf of a selling partner before the selling partner sets the item’s price. The selling partner can then take any estimated fees into account. Each fees estimate request must include an original identifier. This identifier is included in the fees estimate so that you can correlate a fees estimate with the original request. Note: The identifier value is only an estimate, actual costs may vary. Search "fees" in Seller Central and consult the store-specific fee schedule for the most up-to-date information. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 1 | 2 | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation, when available. The table above indicates the default rate and burst values for this operation. Selling partners whose business demands require higher throughput may see higher rate and burst values than those shown here. For more information, see Usage Plans and Rate Limits in the Selling Partner API.